Catch disconnection errors in prepare() too
This may happen for instance if we use server-side preparation. The
trick is that prepare() is already defined in DBConnection, so we reuse
the AUTOLOAD mechanism to call prepare() in a protected manner. But then we
need to be careful *not* to disconnect since otherwise the following
execute() would fail.